// src/types/types.ts
export type ForecastItem = {
  dt: number; // Time of data forecasted, unix, UTC
  main: MainWeatherData; // Main weather data
  weather: WeatherCondition[]; // Array of weather conditions
  clouds: CloudData; // Cloudiness data
  wind: WindData; // Wind data
  visibility: number; // Average visibility, metres
  pop: number; // Probability of precipitation, 0 to 1
  rain?: PrecipitationData; // Rain data (optional)
  snow?: PrecipitationData; // Snow data (optional)
  sys: SystemData; // Part of the day data
  dt_txt: string; // Time of data forecasted, ISO, UTC
};

// Type for the `main` field in `list`
export type MainWeatherData = {
  temp: number; // Temperature
  feels_like: number; // Human-perceived temperature
  temp_min: number; // Minimum temperature
  temp_max: number; // Maximum temperature
  pressure: number; // Atmospheric pressure on sea level, hPa
  sea_level?: number; // Atmospheric pressure on sea level, hPa (optional)
  grnd_level?: number; // Atmospheric pressure on ground level, hPa (optional)
  humidity: number; // Humidity, %
  temp_kf?: number; // Internal parameter (optional)
};

// Type for the `weather` field in `list`
export type WeatherCondition = {
  id: number; // Weather condition id
  main: string; // Group of weather parameters (Rain, Snow, Clouds, etc.)
  description: string; // Weather condition description
  icon: string; // Weather icon id
};

// Type for the `clouds` field in `list`
export type CloudData = {
  all: number; // Cloudiness, %
};

// Type for the `wind` field in `list`
export type WindData = {
  speed: number; // Wind speed
  deg: number; // Wind direction, degrees (meteorological)
  gust?: number; // Wind gust (optional)
};

// Type for the `rain` and `snow` fields in `list`
export type PrecipitationData = {
  "1h"?: number; // Volume for last 1 hour, mm
  "3h"?: number; // Volume for last 3 hours, mm
};

// Type for the `sys` field in `list`
export type SystemData = {
  pod: string; // Part of the day (n - night, d - day)
};
// ...
export interface ItemDetailProps {
  item: ForecastItem;
}